move color data to separate file
